feat: footer에서도 옵저버 구현
